Refactor: Introduce BaseActivity to reduce code duplication
This commit introduces a `BaseActivity` in the `androidMain` source set to encapsulate common logic previously duplicated in the `MainActivity` of different product flavors.

The common logic, including UI state handling, splash screen management, and edge-to-edge theming, has been moved from the `androidFossReliant` and `androidGooglePlay` `MainActivity` files into the new `BaseActivity`. Both flavor-specific `MainActivity` classes now inherit from this base class, simplifying their implementation.
